Реализация 3 модуля. Доработать: profit и отображение фотографий

+++ УП 2 по ТРПО Сервис частных объявлений/Ad_Statuses.cs
@@ -20,7 +20,7 @@ namespace УП_2_по_ТРПО_Сервис_частных_объявлений
             this.Ads = new HashSet<Ads>();
         }
     
-        public double status_id { get; set; }
+        public int status_id { get; set; }
         public string status_name { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
