+++ УП 2 по ТРПО Сервис частных объявлений/Cities.cs
@@ -20,7 +20,7 @@ namespace УП_2_по_ТРПО_Сервис_частных_объявлений
             this.Ads = new HashSet<Ads>();
         }
     
-        public double city_id { get; set; }
+        public int city_id { get; set; }
         public string city_name { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
